Removed the unneeded abstraction for the `MiddlewareEngine`
We really didn't need this abstracted out in a way where you could create multiple passes like we had. There will be very few passes built into GULC and GULC won't be a library for others to extend in that way so there really isn't any reason to abstract it out. We can call the individual passes in the order they need to be called in and end it there.