Modified the `IdentifierExpr` resolver to now say what the identifier is resolved to
`IdentifierExpr` is now replaced in the `DeclResolver` with a new `Expr` based on how it was resolved. If it is a local variable reference it is replaced with `RefLocalVariableExpr`, if it is a function within the file then it is a `RefFileFunctionExpr`, etc. This is intended to make `CodeGen`'s job much easier by not requiring the `CodeGen` to have to search for what the `IdentifierExpr` is again (since it was already found in the `DeclResolver`)